requirements.txt
asgiref==3.6.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
debugpy==1.6.7
decorator==5.1.1
Django==4.1.7
django-js-asset==2.0.0
django-mptt==0.14.0
django-use-email-as-username==1.3.0
djangorestframework==3.14.0
executing==1.2.0
idna==3.4
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.5
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==306
pyzmq==25.0.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
tornado==6.3.1
traitlets==5.9.0
tzdata==2022.7
unicode==2.9
urllib3==1.26.15
wcwidth==0.2.6
